requirements.txt
camelot-py==0.11.0
pypdf==3.12.0
geojson==3.0.1
beautifulsoup4==4.10.0
